requirements.txt
langchain==0.0.184
numpy
python-dotenv==1.0.0
streamlit==1.25.0
streamlit_chat
llama-cpp-python==0.1.66
altair<5
tiktoken
requests
watchdog
Pillow
huggingface_hub
transformers
sentence-transformers
protobuf==3.20.0; sys_platform != 'darwin'
protobuf==3.20.0; sys_platform == 'darwin' and platform_machine != 'arm64'
protobuf==3.20.3; sys_platform == 'darwin' and platform_machine == 'arm64'

# uncomment to use huggingface llms
huggingface-hub==0.14.1

# Utilities
urllib3==1.26.6
accelerate
bitsandbytes
click
flask
requests